requirements.txt
pymupdf
google-genai
watchdog
python-dotenv
customtkinter
